fix WithTimeout issues
