+++ test/gdUnit4Test.csproj
@@ -27,7 +27,8 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="$(MicrosoftSdkVersion)" />
-        <PackageReference Include="gdUnit4.api" Version="4.2.0-rc*" />
+        <!--<PackageReference Include="gdUnit4.api" Version="4.2.0-rc*" / -->
+        <ProjectReference Include="..\api\gdUnit4Api.csproj" />
         <!--ProjectReference Include="gdUnit4.test.adapter" Version="1.0.0-rc*"/-->
         <ProjectReference Include="..\testadapter\gdUnit4TestAdapter.csproj" />
     </ItemGroup>
